Correct SQL Short return

+++ src/main/java/org/kissweb/database/Cursor.java
@@ -320,7 +320,7 @@ public class Cursor implements AutoCloseable {
     }
 
     /**
-     * Return the <code>Integer</code> value of the named column.
+     * Return the <code>Short</code> value of the named column.
      * A <code>null</code> is returned on <code>null</code> valued columns.
      *
      * @param cname
@@ -329,8 +329,8 @@ public class Cursor implements AutoCloseable {
      *
      * @see Record#getShort(String)
      */
-    public Integer getShort(String cname) throws SQLException {
-        return (Integer) get(cname);
+    public Short getShort(String cname) throws SQLException {
+        return (Short) get(cname);
     }
 
     /**
